requirements.txt
python-dotenv==1.0.0
py-cord==2.6.1
pandas==2.3.3
polars==1.34.0
scikit-learn==1.7.2
openai==2.1.0